Split controllers and layouts to own folders
