+++ src/layouts/QueryStringNotPassed.jsx
@@ -3,7 +3,7 @@ import { Card } from "react-bootstrap";
 
 import NotFoundImg from "../assets/undraw_page_not_found_su7k.svg";
 
-const QueryStringNotPasses = ({ lang, translate }) => {
+const QueryStringNotPassed = ({ lang, translate }) => {
 	return (
 		<Card border='light' style={{ width: "100%", padding: "2rem 1rem" }}>
 			<Card.Img variant='top' alt='Not Found' src={NotFoundImg} />
@@ -22,4 +22,4 @@ const QueryStringNotPasses = ({ lang, translate }) => {
 	);
 };
 
-export default QueryStringNotPasses;
+export default QueryStringNotPassed;
